[TableGen] Extract InstructionEncoding class into a separate file (NFC)
